refactor: reorganize scanner feature
