+++ src/drive/web/modules/drive/Toolbar/components/Scanner/Scanner.spec.tsx
@@ -1,0 +1,9 @@
+// The Scanner Feature component should:
+// - render nothing if the folder is not on the mobile app
+// - render nothing if the folder is on the mobile app but the scanner is not available
+// - render an ActionMenuItem if the folder is on the mobile app and the scanner is available
+// - call the startScanner function on click
+// - update the document file after the scan
+// - handle unexpected errors
+
+export {}
