+++ src/drive/web/modules/drive/Toolbar/components/Scanner/ScannerContext.ts
@@ -1,0 +1,9 @@
+import React from 'react'
+
+/**
+ * Context object for the Scanner component.
+ */
+export const ScannerContext = React.createContext({
+  startScanner: () => Promise.resolve(),
+  hasScanner: false
+})
