+++ src/drive/web/modules/drive/Toolbar/components/Scanner/ScannerMenuItem.tsx
@@ -5,18 +5,24 @@ import { ActionMenuItem } from 'cozy-ui/transpiled/react/deprecated/ActionMenu'
 import Icon from 'cozy-ui/transpiled/react/Icon'
 import CameraIcon from 'cozy-ui/transpiled/react/Icons/Camera'
 
-import { ScannerContext } from 'drive/web/modules/drive/Toolbar/components/ScanWrapper'
+import { ScannerContext } from './ScannerContext'
 
-export const StartScanner = (): JSX.Element | null => {
+/**
+ * Renders a scanner menu item.
+ * @returns The JSX element representing the scanner menu item.
+ */
+export const ScannerMenuItem = (): JSX.Element => {
   const { t } = useI18n()
   const scannerContext = useContext(ScannerContext)
 
-  return (
+  return scannerContext.hasScanner ? (
     <ActionMenuItem
       onClick={scannerContext?.startScanner}
       left={<Icon icon={CameraIcon} />}
     >
       {t('Scan.scan_a_doc')}
     </ActionMenuItem>
+  ) : (
+    <></>
   )
 }
